+++ src/command/graph/publish.rs
@@ -14,6 +14,7 @@ use serde::Serialize;
 
 use crate::{
     RoverError, RoverOutput, RoverResult,
+    command::CliOutput,
     options::{CheckConfigOpts, GraphRefOpt, ProfileOpt, SchemaOpt},
     utils::client::StudioClientConfig,
 };
@@ -88,11 +89,17 @@ impl Publish {
             .await
             {
                 Ok(check_res) => {
-                    eprintln!("{}", check_res.get_output());
+                    eprintln!(
+                        "{}",
+                        crate::command::check_output::CheckWorkflowOutput(&check_res).text()
+                    );
                     eprintln!("{}", Style::Success.paint("Check passed. Publishing SDL"));
                 }
                 Err(RoverClientError::CheckWorkflowFailure { check_response, .. }) => {
-                    eprintln!("{}", check_response.get_output());
+                    eprintln!(
+                        "{}",
+                        crate::command::check_output::CheckWorkflowOutput(&check_response).text()
+                    );
                     eprintln!(
                         "{}",
                         Style::Failure.paint(